Split JS code into per-scenario files
Refactor UI files keeping all element attributes on the same line.
