+++ resources/session-finish.view
@@ -1,13 +1,9 @@
 <svg id="background" >
   <rect x="0"   y="0" width="100%" height="100%" fill="#000000" />
-  <animate attributeName="opacity" from="0" to="1" begin="load" dur="0.5"
-  easing="ease-in-out" />
+  <animate attributeName="opacity" from="0" to="1" begin="load" dur="0.5" easing="ease-in-out" />
   <textarea id="description" class="textarea-msg">
-    Are you sure you want to finish this session? Backswipe to cancel. Notice
-    how the previous view hasn't actually been unloaded yet.
+    Are you sure you want to finish this session? Backswipe to cancel. Notice how the previous view hasn"t actually been unloaded meanwhile.
   </textarea>
-  <use id="btn-finish" href="#text-button" class="text-button primary
-  application-fill" />
-  <use id="btn-cancel" href="#text-button" class="text-button secondary
-  application-fill" />
+  <use id="btn-finish" href="#text-button" class="text-button primary application-fill" />
+  <use id="btn-cancel" href="#text-button" class="text-button secondary application-fill" />
 </svg>
