+++ resources/session.view
@@ -1,10 +1,8 @@
 <svg id="background" >
   <rect x="0"   y="0" width="100%" height="100%" fill="#000000" />
-  <animate attributeName="x" to="0" begin="enable" dur="0.5"
-  easing="ease-in-out" />
+  <animate attributeName="x" to="0" begin="enable" dur="0.5" easing="ease-in-out" />
   <textarea id="description" class="textarea-msg">
-    This is an app session simulation. The backswipe gesture is overridden if a
-    session is active, taking the user to an action confirmation view.
+    This is an app session simulation. The backswipe gesture is overridden if a session is active, taking the user to an action confirmation view.
   </textarea>
   <text id="duration" class="duration-msg defaultText" />
   <svg class="container application-fill">
@@ -15,8 +13,7 @@
     </svg>
     <svg class="row">
       <text id="label-animate" class="defaultText" />
-      <use id="animate-toggle" href="#toggle-switch" class="toggle-switch"
-      value="1" />
+      <use id="animate-toggle" href="#toggle-switch" class="toggle-switch" value="1" />
       <rect class="line" />
     </svg>
   </svg>
