+++ resources/single.view
@@ -1,9 +1,7 @@
 <svg id="background" >
   <rect x="0"   y="0" width="100%" height="100%" fill="#000000" />
-  <animate attributeName="x" from="100%" to="0" begin="load" dur="0.5"
-  easing="ease-in-out" />
+  <animate attributeName="x" from="100%" to="0" begin="load" dur="0.5" easing="ease-in-out" />
   <textarea id="description" class="textarea-msg">
-    This is a dummy view, there's nothing here waiting for you! Backswipe to try
-    something else from the main menu.
+    This is a dummy view, there"s nothing here waiting for you! Backswipe to try something else from the main menu.
   </textarea>
 </svg>
